Rewrite to refer to templates for setup
An explanation on how to write tests without templates turns out to be tricky to maintain: I couldn't get the Jest
setup to work following these instructions, and the vitest setup is currently not even working in template form.

The new philosophy for these docs is to defer to templates for the detailed setup work. Only
actual working templates are referred to. So, we have currently no instructions for setting it up
without TypeScript, nor with vitest, as we don't have working templates for this. (there's a vitest
template that is broken in the same way as these instructions - I ran into this as well).