Document workaround for TS error when updating stores
When updating a store, TypeScript complains that `undefined` is not the same type as the value the store holds.

The workaround is to use a non-null assertion operator. I suspect most developers have never asserted `undefined` as non-null before and wouldn't think to do that, so this PR documents that as the solution.